Logic adapter now takes two parameters.
The logic adapter now takes two parameters.
The first is a statement string. The second is a list of strings
that may contain possible matches.

This change simplifies testing and makes it easier to write logic
adapters. This change also prevents other adapters from being
needing to be passed to each other.