Logic adapters should raise an exception if a list of options in not provided.
This adds a new exception class to be used in these cases.
The reason for making this change is because previously,
logic adapters returned the input statement in the case that
the list of possible options is empty. This gives a false
positive that an exact match to the statement exists in the
list. Instead, an exception should be thrown.